chart view

+++ code/frontend/ios-app/OptiFit/OptiFit/Models/Workout/WorkoutSummary.swift
@@ -8,7 +8,7 @@
 import Foundation
 
 struct WorkoutSummary: Codable, Equatable, Hashable {
-    let totalTime: Int
+    let totalTime: Double
     let totalSets: Int
     let totalReps: Int
     let totalWeight: Decimal
